+++ internal/test/recorder.go
@@ -1,0 +1,42 @@
+package test
+
+import (
+	"net/http"
+	"net/http/httptest"
+)
+
+// compatibleResponseRecorder wraps ResponseRecorder to ensure consistent behavior
+// between different versions of Go.
+//
+// Unfortunately, there was a behavior change in go1.6:
+//
+// "The net/http/httptest package's ResponseRecorder now initializes a default
+// Content-Type header using the same content-sniffing algorithm as in
+// http.Server."
+type compatibleResponseRecorder struct {
+	*httptest.ResponseRecorder
+	wroteHeader bool
+}
+
+func newCompatibleResponseRecorder() *compatibleResponseRecorder {
+	return &compatibleResponseRecorder{
+		ResponseRecorder: httptest.NewRecorder(),
+	}
+}
+
+func (rw *compatibleResponseRecorder) Header() http.Header {
+	return rw.ResponseRecorder.Header()
+}
+
+func (rw *compatibleResponseRecorder) Write(buf []byte) (int, error) {
+	if !rw.wroteHeader {
+		rw.WriteHeader(200)
+		rw.wroteHeader = true
+	}
+	return rw.ResponseRecorder.Write(buf)
+}
+
+func (rw *compatibleResponseRecorder) WriteHeader(code int) {
+	rw.wroteHeader = true
+	rw.ResponseRecorder.WriteHeader(code)
+}