+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## ChangeLog
 
+* Fix unit tests broken in 1.6.
+
 * In `Development` mode, the license must be the proper length or empty.
 
 ## 0.5.0